Create Stats
What I want from this program is for the user to be able to enter multiple sets of data. The obstacle is that when the user types 'n' for no more data sets, the program will not remember the values that I have assigned the variables in the function. Data that I want calculated from all data sets is avg, stdv, coeff of variability ect...