Here is the second much more functional version
I added a loop in the program that will give stats on all of the data. Additionally, I made it more lab friendly with specific inputs in the sequence the data would be entered. The next thing I want to figure out for this program is how to get it to yield an excel file with all of the data and calculations once all of the values have been entered.